% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_auc.R
\name{calculate_auc}
\alias{calculate_auc}
\title{calculates the area under the curve for the given model}
\usage{
calculate_auc(model, data, .progress = FALSE)
}
\arguments{
\item{model}{A model generated by the function \code{\link[=mv_model_cox]{mv_model_cox()}} or \code{\link[=mv_model_logreg]{mv_model_logreg()}}}

\item{data}{Multiple imputed data organized as a long dataset}

\item{.progress}{\code{TRUE} to render the progress bar \code{FALSE} otherwise}
}
\value{
The same \code{model} passed as a parameter with the Harrell C Index value stored in \verb{$auc}
}
\description{
This function calculates the Area Under the Curve value for the given model. It uses the \code{\link[pROC:auc]{pROC::auc()}} function from the \code{pROC} package.
}
