% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_brier_score.R
\name{calculate_brier_score}
\alias{calculate_brier_score}
\title{calculate the brier score for the given model}
\usage{
calculate_brier_score(
  model,
  data,
  type = c("prediction", "prediction_type_1", "prediction_type_2"),
  n_boot = 1000,
  seed = NULL
)
}
\arguments{
\item{model}{Model generated with \code{\link[=mv_model_cox]{mv_model_cox()}} or \code{\link[=mv_model_logreg]{mv_model_logreg()}}. Needs the expected prediction parameter already calculated in the model. To generate the predictions you must use the function/s \code{\link[=calculate_predictions]{calculate_predictions()}}, \code{\link[=calculate_predictions_recalibrated_type_1]{calculate_predictions_recalibrated_type_1()}} or \code{\link[=calculate_predictions_recalibrated_type_2]{calculate_predictions_recalibrated_type_2()}}}

\item{data}{Data for what the observed predictions will be calculated.}

\item{type}{Type of the predictions that the calibration plot data should be generated from: \code{"prediction"}, \code{"prediction_type_1"} or \code{"prediction_type_2"}}

\item{n_boot}{number of bootstrap resamples to calculate the Brier Score standar error.}

\item{seed}{random seed generator}
}
\description{
The Brier Score is calculated using the formula \deqn{BS=\frac{1}{n}\sum_{t=1}^n{(f_t - o_t)^2}} where \eqn{n} is the population size, \eqn{f_t} is the predictions for the row \eqn{t} and \eqn{o_t} is the dichotomous observation for the row \eqn{t}.

The function operates according to the chosen \code{model}: if the model is Cox, the survival variable is converted into a binary value (1 if the event occurred during follow-up, 0 otherwise), whereas for a logistic regression model, the dependent variable may already be binary or represent survival time; in the latter case, it is transformed into a binary variable using the same rule as in the Cox model.

The confidence interval is calculated by bootstrap resamples.
}
\examples{
\dontrun{
model |>
  calculate_brier_score(data, type = "predictions_aggregated")
}
}
