% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_predictions.R
\name{calculate_predictions}
\alias{calculate_predictions}
\title{Calculate the predictions for a model}
\usage{
calculate_predictions(model, data)
}
\arguments{
\item{model}{A model generated by the function \code{\link[=mv_model_cox]{mv_model_cox()}} or \code{\link[=mv_model_logreg]{mv_model_logreg()}}}

\item{data}{Data parameter for \code{\link[=calculate_predictions.cox]{calculate_predictions.cox()}} or \code{\link[=calculate_predictions.logreg]{calculate_predictions.logreg()}}}
}
\value{
A model with the parameters \code{predictions_aggregated}, \code{predictions_data}, \code{betax} and \code{betax_data} populated.
}
\description{
A generic method for calculating predictions for a given model. Further parameters must be passed to the S3 methods of each class (\code{\link[=calculate_predictions.cox]{calculate_predictions.cox()}} and \code{\link[=calculate_predictions.logreg]{calculate_predictions.logreg()}})
}
\examples{
set.seed(123)

model <- mv_model_logreg(
  coefficients = list(x = 0.5, z = 0.3),
  formula = event ~ x + z,
  intercept = 1.2
)

data <- data.frame(
  .imp = c(1, 1, 1, 2, 2, 2, 3, 3, 3),
  id = c(1, 2, 3, 1, 2, 3, 1, 2, 3),
  x = rnorm(9, 1, 0.25),
  z = rnorm(9, 2, 0.75)
)


model |>
  calculate_predictions(data)
}
