% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_predictions_recalibrated_type_1.cox.R
\name{calculate_predictions_recalibrated_type_1.cox}
\alias{calculate_predictions_recalibrated_type_1.cox}
\title{Calculates the type 1 recalibrated predictions}
\usage{
\method{calculate_predictions_recalibrated_type_1}{cox}(model, data, .progress = FALSE)
}
\arguments{
\item{model}{Model generated with \code{\link[=mv_model_cox]{mv_model_cox()}}. Needs the \code{predictions} parameter of the model, to generate it the function \code{\link[=calculate_predictions]{calculate_predictions()}} must be executed over the model.}

\item{data}{External validation data. Multiple imputation dataset in long format.}

\item{.progress}{\code{TRUE} to render the progress bar \code{FALSE} otherwise.}
}
\value{
A model with the parameter \code{predictions_recal_type_1} and \code{alpha} populated.
\itemize{
\item \code{predictions_recal_type_1}: stores the type 1 recalibrated predictions stored as follows\tabular{lc}{
   id \tab prediction \cr
   1 \tab 0.03 \cr
   ... \tab ... \cr
   n \tab 0.16 \cr
}

\item \code{alpha}: stores the \eqn{\alpha} recalibration parameter.
}
}
\description{
Using the function \code{\link[=get_recalibrate_params_type_1_cox]{get_recalibrate_params_type_1_cox()}} calculates the recalibration parameters in each of the imputed datasets stored in \code{data}. With all the parameters estimated aggregates them and calculates the recalibrated predictions with these aggregated parameters and the aggregated predictions.
}
\examples{
set.seed(123)

model <- mv_model_cox(
  formula = event ~ 0.5 * (x - 1) + 0.3 * (z - 2),
  S0 = 0.98765
)

data <- data.frame(
  .imp = c(1, 1, 1, 2, 2, 2, 3, 3, 3),
  id = c(1, 2, 3, 1, 2, 3, 1, 2, 3),
  x = rnorm(9, 1, 0.25),
  z = rnorm(9, 2, 0.75),
  event = survival::Surv(rpois(9, 5), rbinom(n = 9, size = 1, prob = 0.5))
)

model |>
  calculate_predictions(data) |>
  calculate_predictions_recalibrated_type_1(data)
}
