% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_predictions_recalibrated_type_1.logreg.R
\name{calculate_predictions_recalibrated_type_1.logreg}
\alias{calculate_predictions_recalibrated_type_1.logreg}
\title{Calculates the type 1 recalibrated predictions for a logistic regression model.}
\usage{
\method{calculate_predictions_recalibrated_type_1}{logreg}(model, data, .progress = FALSE)
}
\arguments{
\item{model}{Model generated with \code{\link[=mv_model_logreg]{mv_model_logreg()}}. Needs the \code{predictions} parameter of the model, to generate it the function \code{calculate_predictions} must be executed over the model. This attribute must be generated using \code{\link[=calculate_predictions]{calculate_predictions()}}}

\item{data}{Data for what the predictions must be recalibrated.}

\item{.progress}{\code{TRUE} to render the progress bar, \code{FALSE} otherwise.}
}
\value{
A model with the parameters \code{predictions_recal_type_1} and \code{alpha_type_1} populated.
\itemize{
\item \code{predictions_recal_type_1}: stores the type 1 recalibrated predictions as follows\tabular{lc}{
   id \tab prediction \cr
   1 \tab 0.03 \cr
   ... \tab ... \cr
   n \tab 0.16 \cr
}

\item \code{alpha_type_1}: stores the \eqn{\alpha} recalibration parameter.
}
}
\description{
Calculates the type 1 recalibrated predictions for a logistic regression model. The type 1 recalibration is defined by an \eqn{\alpha} parameter that updates the value of the \code{intercept} (\eqn{\beta_0}) of the model. The log-odds function is rewritten as follows.

\deqn{log(\frac{p}{1 - p}) = \alpha + \beta_0 + \beta_1 \cdot X_1 + \beta_2 \cdot X_2 + \dots + \beta_p \cdot X_p}

Thus, the predictions are updated by adjusting the \code{intercept} value in the model against the external validation data. The \eqn{\alpha} parameter is estimated in each of the imputed datasets by deriving a logistic regression model using the model log-odds as offset. The coefficients in all the models are aggregated using the mean. Using the aggregated parameter and the aggregated log-odds the new predictions are calculated as follows.

\deqn{\frac{1}{1 + e^{(-(\alpha + (\beta \cdot X)))}}}
}
\examples{
set.seed(123)

model <- mv_model_logreg(formula = event ~ 0.5 * x + 0.3 * z - 1.2)

data <- data.frame(
  .imp = c(1, 1, 1, 2, 2, 2, 3, 3, 3),
  id = c(1, 2, 3, 1, 2, 3, 1, 2, 3),
  event = survival::Surv(rpois(9, 5), rbinom(n = 9, size = 1, prob = 0.5)),
  x = rnorm(9, 1, 0.25),
  z = rnorm(9, 2, 0.75)
)

model |>
  calculate_predictions(data) |>
  calculate_predictions_recalibrated_type_1(data)
}
