% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_predictions_recalibrated_type_1.R
\name{calculate_predictions_recalibrated_type_1}
\alias{calculate_predictions_recalibrated_type_1}
\title{Calculates the type 1 recalibration predictions for a model.}
\usage{
calculate_predictions_recalibrated_type_1(model, data, .progress = FALSE)
}
\arguments{
\item{model}{Model for which the recalibrated predictions are calculated}

\item{data}{Data parameter for \code{\link[=calculate_predictions_recalibrated_type_1.cox]{calculate_predictions_recalibrated_type_1.cox()}} function or for \code{\link[=calculate_predictions_recalibrated_type_1.logreg]{calculate_predictions_recalibrated_type_1.logreg()}} function.}

\item{.progress}{.progress parameter for \code{\link[=calculate_predictions_recalibrated_type_1.cox]{calculate_predictions_recalibrated_type_1.cox()}} function or for \code{\link[=calculate_predictions_recalibrated_type_1.logreg]{calculate_predictions_recalibrated_type_1.logreg()}} function.}
}
\value{
A model with the parameter \code{predictions_recal_type_1} and also the recalibration parameters are populated.
}
\description{
A generic method for calculating the type 1 recalibration predictions for a model.
}
\examples{
set.seed(123)

model <- mv_model_logreg(
  coefficients = list(x = 0.5, z = 0.3),
  formula = event ~ x + z,
  intercept = 1.2
)

data <- data.frame(
  .imp = c(1,1,1,2,2,2,3,3,3),
  id = c(1,2,3,1,2,3,1,2,3),
  x = rnorm(9, 1, 0.25),
  z = rnorm(9, 2, 0.75),
  status = c(1,0,0,1,0,0,1,0,0),
  time = c(2,3,5,2,3,5,2,3,5)
)
data$event <- survival::Surv(data$time, data$status)

model |>
  calculate_predictions(data) |>
  calculate_predictions_recalibrated_type_1(data)
}
