% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_c_index_forestplot.R
\name{get_c_index_forestplot}
\alias{get_c_index_forestplot}
\title{Generates a forestplot to visualize C-Index of different models}
\usage{
get_c_index_forestplot(...)
}
\arguments{
\item{...}{The list of models that will conform the forestplot. If they are passed as variables, the variable name is assigned as model name in the forestplot. But, If they are passed as named arguments the argument name is used in the forestplot}
}
\value{
A forestplot with the C-Index data of the list of models
}
\description{
From a list of models passed as parameters of the function extracts their C-Index value and generates a forestplot to compare their results.
}
\examples{
\dontrun{
get_c_index_forestplot(A = model1, B = model2, C = model3, D = model4)
get_c_index_forestplot(model1, model2, model3, model4)
}
}
