% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_error_message_calculate.R
\name{get_error_message_calculate}
\alias{get_error_message_calculate}
\title{Logic for checkin all the needed parameters for \code{calculate_} functions are
correct}
\usage{
get_error_message_calculate(model, data)
}
\arguments{
\item{model}{\code{MiceExtVal} model to check the errors}

\item{data}{External validation dataset where the model should be evaluated}
}
\value{
\code{NULL} if no errors were found or the actual error if some errors
were found.
}
\description{
Logic for checkin all the needed parameters for \code{calculate_} functions are
correct
}
