% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stratified_calibration_plot.R
\name{get_stratified_calibration_plot}
\alias{get_stratified_calibration_plot}
\title{generates an stratified calibration plot}
\usage{
get_stratified_calibration_plot(
  data,
  n_groups,
  type = "predictions_aggregated",
  ...
)
}
\arguments{
\item{data}{dataset where the dependent variable is}

\item{n_groups}{number of points that should be displayed}

\item{type}{Type of the predictions that the calibration plot data should be generated from: \code{"predictions_aggregated"}, \code{"predictions_recal_type_1"} or \code{"predictions_recal_type_2"}}

\item{...}{models that should be plotted in the stratified calibration plot. If they are a named paramter the name is used as strat}
}
\value{
the stratified calibration plot for the given models
}
\description{
generates an stratified calibration plot
}
\examples{

\dontrun{
get_stratified_calibration_plot(data, 10, model1, model2)
get_stratified_calibration_plot(data, 10, strat1 = model1, strat2 = model2)
}
}
