% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiceExtVal-package.R
\docType{package}
\name{MiceExtVal-package}
\alias{MiceExtVal}
\alias{MiceExtVal-package}
\title{MiceExtVal}
\description{
This package allows the user to externally validate Cox and Logistic Regression models from the data that are usually reported in the articles. The package is formed by functions that allow the user to recalibrate the models and also obtain the discrimination and calibration performance while working with multiple imputed data.

To learn more about MiceExtVal, we recommend you to start from the webpage article

\url{https://metodologianavarrabiomed.github.io/mice-ext-val/articles/get-started.html}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://metodologianavarrabiomed.github.io/mice-ext-val}
  \item Report bugs at \url{https://github.com/metodologianavarrabiomed/mice-ext-val/issues}
}

}
\author{
\strong{Maintainer}: Asier Ballesteros Domínguez \email{asier.ballesteros.dominguez@navarra.es} (\href{https://orcid.org/0000-0002-2195-4669}{ORCID})

}
\keyword{internal}
