% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_model_cox.R
\name{mv_model_cox}
\alias{mv_model_cox}
\title{Creates a cox model.}
\usage{
mv_model_cox(coefficients, means, formula, S0 = NULL)
}
\arguments{
\item{coefficients}{\eqn{\beta} values of the model.}

\item{means}{Mean values of the variables in the derivation dataset.}

\item{formula}{Formula of the model.}

\item{S0}{Value of the \eqn{S_0(t)} function for the time of study.}
}
\value{
A model to be used along the package with the next characteristics that could be empty and will be generated with some other functions in the package.
\itemize{
\item \code{coefficients}: \eqn{\beta} values of the model.
\item \code{means}: Mean values of the variables in the derivation dataset.
\item \code{formula}: Formula of the model.
\item \code{S0}: Value of the \eqn{S_0(t)} function for the time of study.
\item \code{alpha}: Recalibration parameter for the type 1 recalibration.
\item \code{S0_type_2}: Value of the \eqn{S_0(t)} function for the time of study for the type 2 recalibration.
\item \code{beta_overall}: Recalibration parameter for the type 2 recalibration.
\item \code{predictions_aggregated}: Aggregated predictions for the validation data.
\item \code{predictions_data}: All predictions for the validation data including all the imputations.
\item \code{betax}: Aggregated \eqn{\beta \cdot X} values for the validation data.
\item \code{betax_data}: All predictions for the validation data including all the imputations.
\item \code{predictions_recal_type_1}: Aggregated predictions after recalibrating them with type 1 recalibration.
\item \code{predictions_recal_type_2}: Aggregated predictions after recalibrating them with type 2 recalibration.
\item \code{c_index}: Harrell C-Index of the predictions in the validation cohort.
}
}
\description{
Creates a Cox model with all the variables needed to be used in the functions along this package. The Cox model follows

\deqn{S_0(t)^{exp(\beta \cdot X)}}

so we need to fullfill all the requirements of the model. When we are predicting the survival/risk for a \eqn{X} we have to center the values with the meas on the derivation data, normally reported inside the articles.
}
\examples{
model <- mv_model_cox(
  coefficients = list(x = 0.5, z = 0.3),
  means = list(x = 3, z = 0.2),
  formula = event ~ x + z,
  S0 = 0.98765
)

}
