% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_model_logreg.R
\name{mv_model_logreg}
\alias{mv_model_logreg}
\title{Creates a logistic regression model}
\usage{
mv_model_logreg(formula)
}
\arguments{
\item{formula}{The model formula to calculate the linear predictor, including coefficients and intercept}
}
\value{
A model to be used along the package with the next characteristics that could be empty and will be generated with some other functions in the package.
\itemize{
\item \code{formula}: Formula of the model containing the coefficients and the intercept.
\item \code{predictions_imp}: \code{tibble} with the predictions for each of the imputed datasets.
\item \code{predictions_agg}: \code{tibble} with the aggregated predictions for each patient.
\item \code{recal_parameters}: \code{tibble} with the recalibration parameters needed.
\item \code{results_imp}: \code{tibble} with the results in each of the imputations.
\item \code{results_agg}: \code{tibble} with the aggregated results.
}
}
\description{
Creates a logistic regression model with all the requirements for the external validation. The logistic regression model calculates the predictions following the next equation.

\deqn{p(X) = \frac{1}{1 + e^{-(\beta \cdot X)}}}

where we can observe that the prediction only depends on the \eqn{\beta} coefficients and the covariable values, \eqn{X}. The logistic regression model have a parameter called \code{intercept} that usually is represented as \eqn{\beta_0}. The risk prediction is derived from the log-odds function.

\deqn{log(\frac{p}{1 - p}) = \beta_0 + \beta_1 \cdot X_1 + \beta_2 \cdot X_2 + \dots + \beta_p \cdot X_p}

where we can see that the \code{intercept} value is not associated with any covariable. Therefore this function parameters are the minimum needed to obtain the predictions.
}
\examples{
model <- mv_model_logreg(formula = event ~ 0.5 * x + 0.3 * z - 1.2)
}
