% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.MiceExtVal.R
\name{print.MiceExtVal}
\alias{print.MiceExtVal}
\title{A generic function to print the \code{MiceExtVal} model}
\usage{
\method{print}{MiceExtVal}(x, ...)
}
\arguments{
\item{x}{a \code{MiceExtVal} model}

\item{...}{ignored and not passed to any function}
}
\value{
the model printed
}
\description{
A generic function to print the \code{MiceExtVal} model
}
\examples{
model_cox <- mv_model_cox(
  coefficients = list(x = 0.5, z = 0.3),
  means = list(x = 1, z = 2),
  formula = event ~ x + z,
  S0 = 0.98765
)

print(model_cox)

model_logreg <- mv_model_logreg(
  coefficients = list(x = 0.5, z = 0.3),
  formula = event ~ x + z,
  intercept = 1.2
)

print(model_logreg)
}
