% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_brier_score.R
\name{calculate_brier_score}
\alias{calculate_brier_score}
\title{calculate the brier score for the given model definition}
\usage{
calculate_brier_score(
  model,
  data,
  type = c("predictions_aggregated", "predictions_recal_type_1",
    "predictions_recal_type_2"),
  n_boot = 1000,
  seed = NULL
)
}
\arguments{
\item{model}{Model generated with \code{\link[=mv_model_cox]{mv_model_cox()}} or \code{\link[=mv_model_logreg]{mv_model_logreg()}}. Needs the expected prediction parameter already calculated in the model. To generate the predictions you must use the function/s \code{\link[=calculate_predictions]{calculate_predictions()}}, \code{\link[=calculate_predictions_recalibrated_type_1]{calculate_predictions_recalibrated_type_1()}} or \code{\link[=calculate_predictions_recalibrated_type_2]{calculate_predictions_recalibrated_type_2()}}}

\item{data}{Data for what the observed predictions will be calculated.}

\item{type}{Type of the predictions that the calibration plot data should be generated from: \code{"predictions_aggregated"}, \code{"predictions_recal_type_1"} or \code{"predictions_recal_type_2"}}

\item{n_boot}{number of bootstrap resamples to calculate the Brier Score standar error.}

\item{seed}{random seed generator}
}
\description{
The Brier Score is calculated using the formula \deqn{BS=\frac{1}{n}\sum_{t=1}^n{(f_t - o_t)^2}} where \eqn{n} is the population size, \eqn{f_t} is the predictions for the row \eqn{t} and \eqn{o_t} is the dichotomous observation for the row \eqn{t}.

The function works accordingly to the \code{model} definition. If the model is defined as Cox, the survival variable is transformed to dichotomous as \code{1} if the event has appeared during the follow up and \code{0} othercase. In the case the model is a logistic regression the dependent variable can be already dichotomous or survival, in the case of the survival variable the dichotomous is defined as in the Cox model.

The confidence interval is calculated by bootstrap resamples.
}
\examples{
\dontrun{
model |>
  calculate_brier_score(data, type = "predictions_aggregated")
}
}
