% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_c_index.R
\name{calculate_c_index}
\alias{calculate_c_index}
\title{Calculates Harrell C-Index for a model}
\usage{
calculate_c_index(model, data, .progress = FALSE)
}
\arguments{
\item{model}{A model generated by the function \code{\link[=mv_model_cox]{mv_model_cox()}} or \code{\link[=mv_model_logreg]{mv_model_logreg()}}}

\item{data}{Multiple imputation data organized as a long dataset}

\item{.progress}{\code{TRUE} to render the progress bar \code{FALSE} otherwise}
}
\value{
The same \code{model} passed as a parameter with the Harrell C Index value stored in \verb{$c_index}
}
\description{
This function calculates the Harrell C Index value of a \code{model} in each of the imputed dataset stored in \code{data} as a long dataset. The Harrell's C Index calculation is realized using the \code{Hmisc::rcorr.cens} function of the Harrell's package \code{Hmisc}. After calculating the index in each imputed dataset then it aggregates them using the rubin rules with the \code{psfmi::pool_RR} function of the Heymans' package \code{psfmi}. Finally, it stores the result inside the \code{c_index} attribute of the \code{model}.
}
\examples{
set.seed(123)

model <- mv_model_logreg(formula = event ~ 0.5 * x + 0.3 * z - 1.2)

data <- data.frame(
  .imp = c(1, 1, 1, 2, 2, 2, 3, 3, 3),
  id = c(1, 2, 3, 1, 2, 3, 1, 2, 3),
  x = rnorm(9, 1, 0.25),
  z = rnorm(9, 2, 0.75),
  event = survival::Surv(rpois(9, 5), rbinom(n = 9, size = 1, prob = 0.2))
)

model <- model |>
  calculate_predictions(data) |>
  calculate_c_index(data)
}
\references{
F. E. Harrell, “Harrell Miscellaneous (R package Hmisc version 5.1-1),” R-project.org, Sep. 2023, doi: \url{https://cran.r-project.org/package=Hmisc}.
M. Heymans, “Prediction Model Pooling, Selection and Performance Evaluation Across Multiply Imputed Datasets (R package psfmi version 1.4.0),” R-project.org, Jun. 2023, doi: \url{https://cran.r-project.org/package=psfmi}.
}
