% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_predictions.logreg.R
\name{calculate_predictions.logreg}
\alias{calculate_predictions.logreg}
\title{Calculates the predictions for a logistic regression model}
\usage{
\method{calculate_predictions}{logreg}(model, data)
}
\arguments{
\item{model}{Model generated with \code{\link[=mv_model_logreg]{mv_model_logreg()}}}

\item{data}{External validation data. Multiple imputation dataset in long format.}
}
\value{
A model with the parameters \code{predictions_aggregated}, \code{predictions_data}, \code{betax} and \code{betax_data} populated.
\itemize{
\item \code{predictions_aggregated}, stores the predictions aggregated by the mean.
\item \code{predictions_data}, stores all the predictions in each of the imputed datasets.
\item \code{betax}, stores the \eqn{\beta \cdot X} values aggregated by the mean.
\item \code{betax_data}, stores the \eqn{\beta \cdot X} values in each of the imputed datasets.
}
}
\description{
Calculates the predictions of the given logistic regression model and external validation data. The predictions in a logistic regression model are calculated following the next formula.

\deqn{p(X) = \frac{1}{1 + e^{-(\beta \cdot X)}}}

where \eqn{X} stands for the covariable values and \eqn{\beta} for the coefficient values.
}
\examples{
set.seed(123)

model <- mv_model_logreg(
  formula = event ~ 0.5 * x + 0.3 * z + 1.2
)

data <- data.frame(
  .imp = c(1, 1, 1, 2, 2, 2, 3, 3, 3),
  id = c(1, 2, 3, 1, 2, 3, 1, 2, 3),
  x = rnorm(9, 1, 0.25),
  z = rnorm(9, 2, 0.75)
)


model |>
  calculate_predictions(data)
}
