% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_predictions_recalibrated_type_2.cox.R
\name{calculate_predictions_recalibrated_type_2.cox}
\alias{calculate_predictions_recalibrated_type_2.cox}
\title{Calculates the type 2 recalibrated predictions for a Cox model}
\usage{
\method{calculate_predictions_recalibrated_type_2}{cox}(model, data, .progress = FALSE)
}
\arguments{
\item{model}{Model generated with \code{\link[=mv_model_cox]{mv_model_cox()}}. Needs the \code{predictions} parameter of the model, to generate it the function \code{\link[=calculate_predictions]{calculate_predictions()}} must be executed over the model.}

\item{data}{External validation data. Multiple imputation dataset in long format.}

\item{.progress}{\code{TRUE} to render the progress bar \code{FALSE} otherwise.}
}
\value{
A model with the parameter \code{prediction_type_2} added to \code{predictions_agg} and the parameters \code{S0_type_2} and \code{beta_overall} stored in \code{recal_parameters}
\itemize{
\item \code{predictions_agg}: stores now a new variable \code{prediction_type_2}
\item \code{S0_type_2} and \code{beta_overall} are stored in the \code{recal_parameters}.
}
}
\description{
This function calculates the type 2 recalibrated predictions for a Cox model. To help the recalibration of the model the function \code{\link[=get_recalibrate_params_type_2_cox]{get_recalibrate_params_type_2_cox()}} is defined elsewhere in this package. Using this auxiliar function the recalibration parameters are calculated in each of the imputed datasets stored in \code{data} as a long dataset.

After estimating the recalibration parameters in each of the imputed datasets they are aggregated by their mean to use them to recalibrate the predictions of the model. The type 2 recalibration needs from two parameters \code{S0_type_2} and \code{beta_overall}. These parameters are calculated with the \code{\link[=get_recalibrate_params_type_2_cox]{get_recalibrate_params_type_2_cox()}} function. Once they are estimated, they are aggregated by the mean. Finally with the type 2 recalibration parameters and the aggregated predictions the type 2 recalibrated predictions are calculated.

\deqn{S_{0, \text{type 2}}(t)^{exp(\beta_{overall}(\beta \cdot X))}}

where \eqn{S_{0, \text{type 2}}(t)} is estimated using a Weibull distribution and \eqn{\beta_{overall}} is estimated deriving a Cox model with \eqn{\beta \cdot X} as an unique covariate. Both parameters are estimated using the \code{\link[=get_recalibrate_params_type_2_cox]{get_recalibrate_params_type_2_cox()}} function.
}
\examples{
set.seed(123)

model <- mv_model_cox(
  formula = event ~ 0.5 * (x - 1) + 0.3 * (z - 2),
  S0 = 0.98765
)

data <- data.frame(
  .imp = c(1, 1, 1, 2, 2, 2, 3, 3, 3),
  id = c(1, 2, 3, 1, 2, 3, 1, 2, 3),
  x = rnorm(9, 1, 0.25),
  z = rnorm(9, 2, 0.75),
  event = survival::Surv(rpois(9, 5), rbinom(n = 9, size = 1, prob = 0.5))
)

model |>
  calculate_predictions(data) |>
  calculate_predictions_recalibrated_type_1(data) |>
  calculate_predictions_recalibrated_type_2(data)
}
