% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_calibration_plot.R
\name{get_calibration_plot}
\alias{get_calibration_plot}
\title{Generates the calibration plot}
\usage{
get_calibration_plot(data)
}
\arguments{
\item{data}{Outcome of the \code{\link[=get_calibration_plot_data]{get_calibration_plot_data()}} function.}
}
\value{
ggplot2 object ready to be plotted
}
\description{
Generates the calibration plot from the data needed to print it. This data is generated with the \code{\link[=get_calibration_plot_data]{get_calibration_plot_data()}} function.
}
\examples{
\dontrun{
model |>
   get_calibration_plot_data(data, 10, type = "predictions_aggregated") |>
   get_calibration_plot()
}
}
