% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_error_message_calculate_recalibrated.R
\name{get_error_message_calculate_recalibrated}
\alias{get_error_message_calculate_recalibrated}
\title{generates the error messages for the recalibrated predictions}
\usage{
get_error_message_calculate_recalibrated(model, data)
}
\arguments{
\item{data}{}
}
\value{
the error message for the given model and data
}
\description{
generates the error messages for the recalibrated predictions
}
