% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forestplot_data.R
\name{get_forestplot_data}
\alias{get_forestplot_data}
\title{Generates the forestplot data needed in \code{\link[=get_forestplot]{get_forestplot()}}}
\usage{
get_forestplot_data(strat, type = c("c_index", "auc"), ...)
}
\arguments{
\item{strat}{Name of the strat that the models are part of}

\item{...}{A list of models that should be plotted}
}
\value{
A \code{tibble} with the information needed to generate the forestplot
}
\description{
Generates the forestplot data needed in \code{\link[=get_forestplot]{get_forestplot()}}
}
\examples{
\dontrun{
data <- get_forestplot_data(strat = "overall", cox_model, logreg_model)
data <- get_forestplot_data(strat = "overall", Cox = cox_model, LogReg = logreg_model)
}
}
