% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_model_cox.R
\name{mv_model_cox}
\alias{mv_model_cox}
\title{Creates a cox model.}
\usage{
mv_model_cox(formula, S0 = NULL)
}
\arguments{
\item{formula}{Formula of the model to calculate the \eqn{\beta \cdot X} values, including coefficients and means if needed.}

\item{S0}{Value of the \eqn{S_0(t)} function for the time of study.}
}
\value{
A model to be used along the package with the next characteristics that could be empty and will be generated with some other functions in the package.
\itemize{
\item \code{formula}: Formula of how the \eqn{\beta \cdot X} will be calculated.
\item \code{S0}: Value of the \eqn{S_0(t)} function for the time of study.
\item \code{predictions_imp}: \code{tibble} with the predictions for each of the imputed datasets.
\item \code{predictions_agg}: \code{tibble} with the aggregated predictions for each patient.
\item \code{recal_parameters}: \code{tibble} with the recalibration parameters needed.
\item \code{results_imp}: \code{tibble} with the results in each of the imputations.
\item \code{results_agg}: \code{tibble} with the aggregated results.
}
}
\description{
Creates a Cox model with all the variables needed to be used in the functions along this package. The Cox model follows

\deqn{S_0(t)^{exp(\beta \cdot X)}}

so we need to fullfill all the requirements of the model. When we are predicting the survival/risk for a \eqn{X} we have to center the values with the meas on the derivation data, normally reported inside the articles.
}
\examples{
model <- mv_model_cox(
  formula = event ~ 0.5 * (x - 3) + 0.3 * (z - 0.2),
  S0 = 0.98765
)

}
