% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_model_logreg.R
\name{mv_model_logreg}
\alias{mv_model_logreg}
\title{Creates a logistic regression model}
\usage{
mv_model_logreg(coefficients, formula, intercept)
}
\arguments{
\item{coefficients}{A named list containing the \eqn{\beta} coefficients of the model without the intercept}

\item{formula}{The model formula}

\item{intercept}{Intercept value}
}
\value{
A model to be used along the package with the next characteristics that could be empty and will be generated with some other functions in the package.
\itemize{
\item \code{coefficients}: \eqn{\beta} values of the model without the intercept.
\item \code{formula}: Formula of the model.
\item \code{intercept}: Intercept value of the model.
\item \code{alpha_type_1}: The \eqn{\alpha} value for the type 1 recalibration.
\item \code{alpha_type_2}: The \eqn{\alpha} value for the type 2 recalibration.
\item \code{beta_overall}: The \eqn{\beta_{overall}} value for the type 2 recalibration.
\item \code{predictions_aggregated}: Aggregated predictions for the validation data.
\item \code{predictions_data}: All of the predictions for the validation data in each imputation.
\item \code{betax}: Aggregated \eqn{\beta \cdot X} values for the validation data.
\item \code{betax_data}: All the \eqn{\beta \cdot X} values for the validation in each imputation.
\item \code{predictions_recal_type_1}: Aggregated predictions after recalibrating them with type 1 recalibration.
\item \code{predictions_recal_type_2}: Aggregated predictions after recalibrating them with type 2 recalibration.
\item \code{c_index}: Harrell C-Index of the predictions in the validation cohort.
}
}
\description{
Creates a logistic regression model with all the requirements for the external validation. The logistic regression model calculates the predictions following the next equation.

\deqn{p(X) = \frac{1}{1 + e^{-(\beta \cdot X)}}}

where we can observe that the prediction only depends on the \eqn{\beta} coefficients and the covariable values, \eqn{X}. The logistic regression model have a parameter called \code{intercept} that usually is represented as \eqn{\beta_0}. The risk prediction is derived from the log-odds function.

\deqn{log(\frac{p}{1 - p}) = \beta_0 + \beta_1 \cdot X_1 + \beta_2 \cdot X_2 + \dots + \beta_p \cdot X_p}

where we can see that the \code{intercept} value is not associated with any covariable. Therefore this function parameters are the minimum needed to obtain the predictions.
}
\examples{
model <- mv_model_logreg(
  coefficients = list(x = 0.5, z = 0.3),
  formula = event ~ x + z,
  intercept = 1.2
)
}
